[fix] stm8: Change DwarfRegNum for custom (sub)registers
The LLVM manual states

   -1 is a special Dwarf number that indicates the gcc number is undefined

As we have oriented ourself on the SDCC, which uses these definitions as
dwarf numbers, it makes sense to keep the ones that have already been
defined by SDCC.

But we use a few more (sub)registers (PCE,PCL,PCH) which aren't declared
by SDCC and as such we don't have a dwarf number for them yet. So set
them to -1.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>