[no build] stm8: Add STM8-unknown-unknown target triple
Previously the STM8 target was known to clang (in other words it was
listed under `--print-targets`), but it could not be selected in any
meaningful way because the target triple wasn't properly defined.

  # example failing command
  clang --target=stm8-unknown-unknown main.c

With this change, this has been resolved now and together with the new
`STM8TargetMachine` it is now possible to define the triple
`stm8-unknown-unknown` as a valid target.

Note
The `STM8TargetMachine` defines parts of the data layout of STM8. The
data layout chosen there has been taken from the data layout of SDCC.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>