[fix] stm8: Pointer alignment in datalayout string
After compiling llvm it complained at runtime, that the data-layout
string which we introduced in one of the recent commits did not stand up
to its standards. More specifically we only defined the size of the
pointer type, but not the alignment.

This commit fixes that and additionally adds the datalayout string to
the `STM8TargetMachine` file to replace the previous `TODO` placeholder.

Note: There are currently two versions of the datalayout string that are
identical. This is similarly done in all other targets without us
knowing why.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>