[wip] stm8: register target register info
This solves the following assertion being triggered during run of clang

  Assertion `MRI && "Unable to create reg info"

The solution was rather simple: The registry happens in the
`STM8MCTargetDesc` file in `LLVMInitializeSTM8TargetMC`. We haven't
put anything there yet, so of course would clang crash very early on in
the process. Luckily we already introduced the register descriptions in
a previous commit.

This is of course a non working sample and further infos must be
registered before clang will be able to do anything.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>