[wip] stm8: register ASM info
Adds a simple description of the STM8 asm code. This is currently not
based on anything already existing and is mostly there for the compiler
to have something to feed on.

The STM8MCAsmInfo initialization should be adapted along the way to
match the ASM format of the SDCC more closely - along interoperability.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>