stm8: Change pointer size for internal LLVM use
During the SelectionDAG stage (specifically the lowering of the LLVM IR
to the target rep.) clang crashed in the very early stages as it
couldn't initialize an `SDValue` with the correct value type.

This was due to the fact, that `ValueTypes.td` does not define an i24
type, which we previously required for the STM8 pointer type as it is
described in the datalayout string.

The trivial sounding solution would be to add the i24 data type to the
value types, but this is heavily discouraged as it would break binary
compatibility with previous LLVM versions.[^1]

Luckily we can override the transformation 'bit depth -> IR value type'
with our own implementation in `TargetLowering` by adapting the method
`getPointerTy` which handles these transformations normally. The adapted
method will either call the previous implementation or (in case of the
i24), replace it with the next biggest (and supported) integer type i32.

[^1]: https://llvm.org/docs/ExtendingLLVM.html#adding-a-new-type

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>