[WIP] stm8: Add STM8CallingConv.td
Using the calling convention by SDCC for STM8 as a reference, this is a
first step in implementing this calling convention in LLVM.

The above mentioned `STM8CallingConv.td` is responsible for handling the
easy aspect of that like in our case the return calling conv or varargs.

As I've never written this type of file for LLVM before, it might might
be subject to changes.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>