[wip] stm8: Improve STM8RegistersInfo class
This class was previously only a stub and as such did not implement any
useful changes nor did it implement everything that was required for it
not to be an abstract class.

With this all at least required methods are implement as stubs, and the
trivial ones also have a valid implementation now.

What is notably missing here is:
- elimateFrameIndex: this is something I haven't yet bothered to
  understand. But according to AVR it does something with a stack I
  guess
- getPointerRegClass: we have different pointer sizes regarding the
  addressing mode. This might have an influence here

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>