stm8: Introduce STM8 calling convention to LLVM
The current goal is to only implement the current calling convention
used by the open source SDCC compiler (SDCC 4.4.0 STM8 CC version 1 in
the manual), which we will here on out call `STM8_BUILTIN`.

Additionally we need to consider the special situation in which we are
in an ISR on the STM8, which has been accordingly named `STM8_INTR`.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>