[WIP] stm8: Introduce new instr formats
The previous iteration of instruction formats introduced in

> [wip] stm8: Register instruction info + skeleton

was incomplete at best. A deeper analysis of the STM8 mnemonic arguments
showed a few patterns, that are (partially) integrated in this revision.

For now a few possible single operand encodings have their own class:

-
- BYTE
- WORD
- SHORTMEM
- LONGMEM
- SHORTOFF_?
- SHORTPTRW
- LONGPTRW

Missing are:
- LONGOFF_?
- LONGMEM, PCREL
- EXTMEM
- LONGPTRE
- LONGPTRE_?
- PCREL
- EXTOFF_?
- LONGMEM, BYTE
- SHORTMEM, SHORTMEM
- LONGMEM, LONGMEM

As I'm not fully familiar with the TableGen language and especially the
In and Out DAGs yet, I've not integrated all of them as can be argued,
that e.g. `LONGMEM` & `LONGOFF_?` or `SHORTMEM` and `SHORTOFF_?` are
just the same instruction class, but with a different name. Further
development here might show that separating them so finely might prove
useful (for example for custom and unique `EncoderMethod`s).

With this change I've also taken a stab at integrating a single ADD
instruction just to see how the formats do in the wild (and because the
build system requires at least one valid instruction).

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>