[wip] stm8: Describe STM8 target properties for lowering
This is a required step, so that LLVM knows what the target supports,
which instructions, registers, … require additional care and so on.

In our case these are the stack pointer, the alignment, register classes
and quite important the `computeRegisterProperties` step. Without the
latter being called, LLVM will have no simple types available throughout
the process.

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>