STM8: Remove PC register definition
It was noticed, that AVR does not reserve nor define the PC register in
its backend (see [the discussion][dis]). The current theory is, that due
to the fact that one cannot directly interact with the PC register -
except through specialised instructions (jump, …) - it is not necessary
to define it as an register, which would be used internally by LLVM to
describe mnemonic operands.

We also noticed adapting the definitions, that `getPointerRegClass` is
used to describe registers that can hold pointer values, instead of it
describing the PC register, as assumed before. As our ISA contains the
instructions JP (X) or JP (Y), whereby X and Y hold the pointer, the
previous PC reg class was replaced by a class containing these two.

[dis]: https://github.com/nrdmn/llvm-project/pull/5#discussion_r1538079279

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>